requirements.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
PyPDF2==3.0.1
python-docx==0.8.11
plotly==5.15.0
seaborn==0.12.2
matplotlib==3.7.2
streamlit-option-menu==0.3.6
streamlit-lottie==0.0.5